# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AIClient-2-API
  description: |
    AIClient-2-API 是一个将多种大模型 API（Gemini, Claude, OpenAI 等）统一封装为 OpenAI 兼容接口的代理服务。
    
    ## 功能特性
    - 支持多种大模型提供商（Gemini, Claude, OpenAI, Qwen, iFlow, Zhipu 等）
    - OpenAI 兼容接口
    - 多账号轮询和智能故障转移
    - 实时日志推送
    - Web UI 管理界面
    - OAuth 授权机制
    
    ## 认证方式
    - **管理后台 API**：Bearer Token（通过 `/api/login` 获取）
    - **核心 AI API**：
      - `Authorization: Bearer <key>` - OpenAI 风格
      - `x-api-key: <key>` - Claude 风格
      - `x-goog-api-key: <key>` - Gemini 风格
      - URL 参数 `?key=<key>` - Gemini 风格
  version: 1.0.0
  contact:
    name: AIClient-2-API Team
    url: https://github.com/4xiaxia/4x
servers:
  - url: http://localhost:3000
    description: 本地开发服务器
  - url: http://localhost:3100
    description: 主进程管理服务器
...
paths:
...
  /ollama/api/show:
    post:
      tags:
        - Ollama API
      summary: 显示模型详情
      description: Ollama 兼容的模型详情接口
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        '200':
          description: 模型详情获取成功
          content:
            application/json:
              schema:
                type: object

  # ==================== 主进程接口 ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token（通过 /api/login 获取）

    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: API Key（支持多种格式）
      x-api-key: API Key（Claude 风格）
      x-goog-api-key: API Key（Gemini 风格）
...